Move BaseCamera to GameCamera as right now there is no value in a base class. Change magnitudes to floats instead of ints. Add more functionality for continuous movement (so that holding down keys continually moves / zooms rather than incremental).

+++ FrigidRogue.MonoGame.Core/Graphics/Camera/IGameCamera.cs
@@ -7,6 +7,6 @@ namespace FrigidRogue.MonoGame.Core.Graphics.Camera
         CameraMovementType ContinuousCameraMovementType { get; set; }
         void Move(CameraMovementType cameraMovementType, float moveMagnitude);
         void Rotate(CameraMovementType cameraMovementType, float rotateMagnitude);
-        void Zoom(int magnitude);
+        void Zoom(float magnitude);
     }
 }
